requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
altgraph @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.6.3
argcomplete==1.10.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.8.2
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
compressed-rtf==1.0.6
cryptography==41.0.7
distro==1.8.0
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.3
extract-msg==0.28.7
frozenlist==1.4.1
future @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.14.0
httpcore==1.0.1
httpx==0.25.2
idna==3.4
IMAPClient==2.1.0
inflection==0.5.1
lxml==4.9.3
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.2
olefile==0.47
openai==1.3.6
opencv-python==4.8.1.78
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyPDF2==3.0.1
pypdfium2==4.24.0
pyth==0.6.0
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.23
python-telegram-bot==20.7
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
secure-smtplib==0.1.1
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.23
striprtf==0.0.26
textract==1.6.5
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==1.26.18
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.4
